requirements.txt
fastapi==0.112.2
uvicorn[standard]==0.30.6
httpx==0.27.2
sqlalchemy
passlib==1.7.4
bcrypt==4.0.1
python-jose[cryptography]
